+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+### Added
+
+- Refresh customer subscription cache - distribution
+
+### Changed
+
+- Register-usage-record - distribution filter canceled subscription
+
 ## [3.0.0] - 2022-06-29
 
 ### Added